+++ R/meta_packages.R
@@ -40,7 +40,7 @@ meta_packages <- function(
   foss <- utils::available.packages(repos = repo, filters = "license/FOSS")
   out$foss <- FALSE
   out[as.character(foss[, "Package"]), "foss"] <- TRUE
-  freeze <- date_staging()
+  freeze <- meta_snapshot()$staging
   p3m <- "https://packagemanager.posit.co"
   repo_cran <- file.path(p3m, "cran", freeze)
   cran <- utils::available.packages(repos = repo_cran)