+++ R/update_staging.R
@@ -86,7 +86,7 @@ update_staging <- function(
   json_new <- json_new[order(json_new$package), ]
   jsonlite::write_json(json_new, file_staging, pretty = TRUE)
   file_config <- file.path(path_staging, "config.json")
-  json_config <- list(cran_version = date_staging())
+  json_config <- list(cran_version = meta_snapshot()$staging)
   jsonlite::write_json(
     json_config,
     file_config,